requirements.txt
pywin32
anthropic
python-dotenv
customtkinter
pystray
Pillow
keyboard
pyinstaller
sentry-sdk
